requested change demo page
Refs: #6249
