+++ packages/samples/react/src/components/tabs/routes.ts
@@ -1,12 +1,12 @@
 import { Routes } from '../../shares/types';
 import { TabsBasic } from './basic';
 import { TabsIconsOnly } from './icons-only';
-import { TabsBehaviorSelectManual } from './behavior-select-manual';
+import { TabsBehavior } from './behavior';
 
 export const TABS_ROUTES: Routes = {
 	tabs: {
 		basic: TabsBasic,
 		'icons-only': TabsIconsOnly,
-		'behavior-select-manual': TabsBehaviorSelectManual,
+		behavior: TabsBehavior,
 	},
 };
